collector/requirements.txt
pywin32==308
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
pyinstaller==6.11.1
